Polish: extract the Connect redirect module, tighten seams, dedupe helpers
